Machine learning principles
we will see what machine learning is and what are the characteristics of the projects that can be carried out with these techniques. You will learn to differentiate the different learning contexts (supervised, unsupervised and reinforcement learning), as well as their associated tasks. In addition, you will learn about the elements that make up a learning algorithm and the recommended cycle to follow to build data-driven projects. Finally, you will have the opportunity to interact with some tools, in the Python programming language, used for data manipulation.